small warning fix for alpha
2006-09-02  Dalibor Topic  <robilad@kaffe.org>

        * kaffe/kaffe/version.c (printShortVersion): Removed
        unused code.

        * kaffe/kaffevm/gc.h: Include config-int.h.
        Define UNLIMITED_HEAP as INT32_MAX, since it is assigned to
        a jint. That fixes four warnings on alpha-linux.

+++ kaffe/kaffevm/gc.h
@@ -16,6 +16,8 @@
 #include "config.h"
 #endif
 
+#include "config-int.h"
+
 #if defined(HAVE_LIMITS_H)
 #include <limits.h>
 #endif
@@ -25,7 +27,7 @@
 /*
  * Default values for initial and maximum heap size and heap increment.
  */
-#define UNLIMITED_HEAP  (ULONG_MAX)
+#define UNLIMITED_HEAP  (INT32_MAX)
 #define	MIN_HEAPSIZE	(5*1024*1024)
 #define	MAX_HEAPSIZE	(UNLIMITED_HEAP)
 #define	ALLOC_HEAPSIZE	(1024*1024)
